requirements.txt
kivy
kivymd
transformers
aiosqlite
aiohttp
cachetools
opencv-python-headless
scenedetect
onnxruntime-gpu
Pillow
watchdog
cryptography
pypdf
ffmpeg-python
easyocr
python-magic
fastapi[all]
PyJWT 
passlib[bcrypt]
jinja2
colorama
keyring
bcrypt
ImageHash
ftfy 
regex
faiss-cpu
git+https://github.com/openai/CLIP.git
fastapi-limiter